requirements.txt
azure-functions
azure-iot-device
meross_iot